Phase 4 fix (UCEF gate): avg_r numpy-vs-native rounding divergence
run-all-gates verifier found a real 0.01 divergence in avg_r on half-cent means (AMC daily
18 trades -> -0.355; MU daily 1.5/104 4 trades -> -0.265). Cause: backtest.py did
round(np.mean(rs), 2) where np.mean returns numpy.float64, and round(np.float64,2) uses
numpy's decimal rounding — different from native round() (and Rust format), which every
OTHER stat (total_r, per-trade r, win_rate) uses. So only avg_r/expectancy were inconsistent.

Fix: round(float(np.mean(rs)), 2) — cast to native float first so avg_r rounds consistently
with the rest and matches the Rust port. (Also removes a latent note-flip risk at the 0.1
expectancy threshold.)

Re-proof: BACKTEST PARITY ALL PASS across weekly/daily x param variations incl. the two
boundary cases and high-trade-count tickers (AMC 18, PLUG 13). Added
tests/parity/backtest_check.sh driver.

+++ tools/backtest.py
@@ -171,7 +171,10 @@ def backtest_frame(df, market, atr_mult: float = 2.0, max_hold: int = 52,
     timeouts = sum(1 for t in trades if t.outcome == "TIMEOUT")
     rs = [t.r for t in trades]
     total_r = round(sum(rs), 2)
-    avg_r = round(np.mean(rs), 2) if rs else 0.0
+    # cast np.mean's numpy.float64 to a native float FIRST: round(np.float64, 2) uses numpy's
+    # decimal rounding, which differs from native round() (and every other stat here) on
+    # half-cent means (e.g. -0.355). float() makes avg_r round consistently with total_r/r.
+    avg_r = round(float(np.mean(rs)), 2) if rs else 0.0
     win_rate = round(100 * wins / len(trades), 1) if trades else 0.0
     gross_win = sum(r for r in rs if r > 0)
     gross_loss = -sum(r for r in rs if r < 0)
